requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
opencv-python==4.7.0.72
easyocr==1.5.1
numpy==1.23.5
Pillow==9.3.0
python-multipart==0.0.6
re
